XRDDEV-746 Move components related to settings, diagnostics and clients
