+++ src/proxy-ui-api/frontend/src/router.ts
@@ -2,33 +2,33 @@ import Router, { Route } from 'vue-router';
 import { sync } from 'vuex-router-sync';
 import Login from './views/Login.vue';
 import Base from './views/Base.vue';
-import Clients from './views/Clients.vue';
+import Clients from '@/views/Clients/Clients.vue';
+import Client from '@/views/Clients/Client.vue';
 import Keys from './views/Keys.vue';
-import Settings from './views/Settings.vue';
-import Diagnostics from './views/Diagnostics.vue';
+import Settings from '@/views/Settings/Settings.vue';
+import SystemParameters from '@/views/Settings/SystemParameters.vue';
+import BackupAndRestore from '@/views/Settings/BackupAndRestore.vue';
+import Diagnostics from '@/views/Diagnostics/Diagnostics.vue';
 import AddSubsystem from './views/AddSubsystem.vue';
 import AddClient from './views/AddClient.vue';
-import Subsystem from './views/Subsystem.vue';
-import Client from './views/Client.vue';
+import Subsystem from './views/Clients/Subsystem.vue';
 import TabsBase from '@/views/TabsBase.vue';
 import ClientTlsCertificate from '@/views/ClientTlsCertificate.vue';
 import Error from '@/views/Error.vue';
-import ClientDetails from '@/components/ClientDetails.vue';
-import InternalServers from '@/components/InternalServers.vue';
-import LocalGroups from '@/components/LocalGroups.vue';
-import LocalGroup from '@/views/LocalGroup.vue';
-import Services from '@/components/Services.vue';
-import ServiceClients from '@/components/ServiceClients.vue';
+import ClientDetails from '@/views/Clients/ClientDetails.vue';
+import InternalServers from '@/views/Clients/InternalServers.vue';
+import LocalGroups from '@/views/Clients/LocalGroups.vue';
+import LocalGroup from '@/views/LocalGroup/LocalGroup.vue';
+import Services from '@/views/Clients/Services.vue';
+import ServiceClients from '@/views/Clients/ServiceClients.vue';
 import ServiceDescriptionDetails from '@/views/ServiceDescriptionDetails.vue';
 import SignAndAuthKeys from '@/components/SignAndAuthKeys.vue';
 import SSTlsCertificate from '@/components/SSTlsCertificate.vue';
 import ApiKey from '@/components/ApiKey.vue';
-import SystemParameters from '@/components/SystemParameters.vue';
-import BackupAndRestore from '@/components/BackupAndRestore.vue';
 import Token from '@/views/Token.vue';
 import Key from '@/views/Key.vue';
 import Certificate from '@/views/Certificate.vue';
-import Service from '@/views/Service.vue';
+import Service from '@/views/Service/Service.vue';
 import store from './store';
 import { RouteName, Permissions } from '@/global';
