+++ src/proxy-ui-api/frontend/src/views/Clients/AddRestDialog.vue
@@ -94,6 +94,6 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scoped>
-@import '../assets/dialogs';
+@import '../../assets/dialogs';
 </style>
