+++ src/proxy-ui-api/frontend/src/views/Clients/AddWsdlDialog.vue
@@ -77,6 +77,6 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scoped>
-@import '../assets/dialogs';
+@import '../../assets/dialogs';
 </style>
