+++ src/proxy-ui-api/frontend/src/views/Clients/ClientsDataTable.vue
