+++ src/proxy-ui-api/frontend/src/views/Clients/DisableServiceDescDialog.vue
@@ -58,6 +58,6 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scoped>
-@import '../assets/dialogs';
+@import '../../assets/dialogs';
 </style>
