+++ src/proxy-ui-api/frontend/src/views/Clients/NewGroupDialog.vue
@@ -83,6 +83,6 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scoped>
-@import '../assets/dialogs';
+@import '../../assets/dialogs';
 </style>
