+++ src/proxy-ui-api/frontend/src/views/Clients/ServiceClients.vue
