+++ src/proxy-ui-api/frontend/src/views/Clients/Subsystem.vue
