+++ src/proxy-ui-api/frontend/src/views/Diagnostics/Diagnostics.vue
