+++ src/proxy-ui-api/frontend/src/views/Service/AccessRightsDialog.vue
@@ -278,8 +278,8 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scoped>
-@import '../assets/tables';
-@import '../assets/add-dialogs';
+@import '../../assets/tables';
+@import '../../assets/add-dialogs';
 
 .first-column {
   width: 40px;
